Split GlobalStore into 2 files to support fast refresh
